Solve high CPU usage
I was testing this software with a Tritech Micron (btw, thanks for sharing this great work) and I notice a too high CPU usage.
I am not completely aware of implications of adding the `usleep(100);` in the loop but it solves the high CPU usage problem, due maybe a too fast pooling. 
It seems a bit "hacky" can you give some thoughts?